implement password-strength service
