+++ src/app/infrastructure/enums/password-strength.enum.ts
@@ -1,0 +1,8 @@
+
+export const enum PasswordStrength{
+  NoStrength = 'no-strength',
+  TooShort = 'too-short',
+  Easy = 'easy',
+  Medium = 'medium',
+  Strong = 'strong'
+}
