+++ src/app/password-form/password-form.component.html
@@ -1,6 +1,8 @@
 
 <div class="password-area">
   <app-password-control [formControl]="passwordControl"/>
+  <!--  developer temporary addition-->
   <p>Value: {{ passwordControl.value }}</p>
+  <!--                              -->
   <app-indicator [state]="currentPasswordStrength"></app-indicator>
 </div>
